Extract massive methods following Discord.js best practices
Major refactoring to reduce CommandHandler.js complexity:

**Extracted Commands:**
- /stop command (106 lines) → src/commands/stop.js
- /transcribe command (232 lines) → src/commands/transcribe.js

**Key Improvements:**
- Reduced CommandHandler.js from 1,045 to 703 lines (-342 lines total)
- Broke down complex methods into focused, testable functions
- Implemented dependency injection pattern for service access
- Added proper error handling and timeout management
- Maintained all functionality while improving maintainability

**Architecture Benefits:**
- Each command is self-contained with clear responsibilities
- Better separation of concerns and single responsibility principle
- Easier to test individual commands in isolation
- Follows Discord.js community standards for command organization
- Reduced cognitive load when working on specific commands

**Technical Details:**
- Commands with dependencies use wrapper pattern in command loader
- Timeout functionality preserved in extracted commands
- Service injection allows access to voiceRecorder, audioProcessor, etc.
- Background job management maintained for async operations

This brings CommandHandler.js closer to Discord.js best practices while
preserving the advanced background processing architecture.
